Add init script and results

+++ init.sh
@@ -4,17 +4,19 @@ DATA_FILENAME=keys.json
 
 yarn build
 
-mkdir -p data
-echo "Generating keys with login ...";
-yarn -s data > $DATA_FOLDER/$DATA_FILENAME
-
-if [ $? -eq 0 ]
+if [ ! -f $DATA_FOLDER/$DATA_FILENAME]
 then
-  echo "Successfully generate and register test lollipop keys"
-else
-  rm -rf $DATA_FOLDER/$DATA_FILENAME
-  echo "Error generating the test lollipop keys" >&2
-  exit 1
+  mkdir -p data
+  echo "Generating keys with login ...";
+  yarn -s data > $DATA_FOLDER/$DATA_FILENAME
+  if [ $? -eq 0 ]
+  then
+    echo "Successfully generate and register test lollipop keys"
+  else
+    rm -rf $DATA_FOLDER/$DATA_FILENAME
+    echo "Error generating the test lollipop keys" >&2
+    exit 1
+  fi
 fi
 
 echo "Starting signer service to generate lollipop signatures ...";
